Remove allow(dead_code) by splitting apc_builder_utils into its own module
Each integration test binary is a separate crate, so `mod common;`
compiles the entire common module into every binary. The
`apc_builder_utils` sub-module was only used by the 4 `apc_builder_*`
tests but was visible (and flagged as dead code) in `machine_extraction`.

Fix this structurally:
- Move `apc_builder_utils` from an inline sub-module in `common/mod.rs`
  to a standalone file `common/apc_builder_utils.rs`.
- Each `apc_builder_*` test includes it via `#[path]`, so it is only
  compiled into the binaries that use it.
- Remove the unused `compile()` helper.
- `common/mod.rs` is now minimal (only `original_vm_config()`), with no
  `#[allow(dead_code)]` needed.
