+++ openvm-riscv/tests/apc_builder_superblocks.rs
@@ -1,4 +1,7 @@
+#[path = "common/apc_builder_utils.rs"]
+mod apc_builder_utils;
 mod common;
+
 use openvm_instructions::instruction::Instruction;
 use openvm_stark_sdk::p3_baby_bear::BabyBear;
 use powdr_autoprecompiles::blocks::BasicBlock;
@@ -6,7 +9,7 @@ use powdr_openvm_riscv::symbolic_instruction_builder::*;
 use test_log::test;
 
 fn assert_machine_output(program: Vec<BasicBlock<Instruction<BabyBear>>>, test_name: &str) {
-    common::apc_builder_utils::assert_machine_output(program.into(), "superblocks", test_name);
+    apc_builder_utils::assert_machine_output(program.into(), "superblocks", test_name);
 }
 
 fn bb(
